requirements.txt
pandas
matplotlib
seaborn
pycryptodome